Revert "chore: support for es2015 (e.g. legacy webpack, node version) (#14)"
This reverts commit 202c978baa30717db679cd637121670fa8e5598e.
