+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.1",
   "description": "Json-Fusion is a Node.js library designed to streamline the process of loading, merging, and reflecting directory hierarchies in JSON files. With Json-Fusion, you can consolidate information scattered across different JSON files, and reflect your filesystem's structure directly in your data, all with just a few lines of code.",
   "repository": {
-    "url": "https://github.com/hacomono-lib/json-fusion",
+    "url": "git@github.com:hacomono-lib/json-fusion.git",
     "type": "git"
   },
   "author": "mew-ton <mxl_usmarc@slmail.mewton.jp>",
@@ -19,16 +19,6 @@
   "main": "./dist/index.cjs",
   "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    },
-    "./dist/es2015": {
-      "import": "./dist/es2015/index.js",
-      "require": "./dist/es2015/index.cjs"
-    }
-  },
   "scripts": {
     "dev": "vitest",
     "build": "tsup",
@@ -44,13 +34,10 @@
     "globby": "^11.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.22.9",
-    "@babel/node": "^7.22.6",
     "@changesets/changelog-github": "^0.4.8",
     "@changesets/cli": "^2.26.2",
     "@typescript-eslint/eslint-plugin": "^6.1.0",
     "@typescript-eslint/parser": "^6.1.0",
-    "babel-plugin-module-resolver": "^5.0.0",
     "eslint": "^8.45.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
