--- test/runs/.babelrc
@@ -1,10 +1,0 @@
-{
-  "plugins": [
-    ["module-resolver", {
-      "root": ["../../../"],
-      "alias": {
-        "json-fusion": "./"
-      }
-    }]
-  ]
-}
